refactor : 重构数据库字段类型

+++ big-market-api/src/main/java/app/xlog/ggbond/IRaffleApiService.java
@@ -5,8 +5,8 @@ import com.fasterxml.jackson.databind.JsonNode;
 
 public interface IRaffleApiService {
 
-    public Response<JsonNode> queryAwardList(Integer strategyId);
+    public Response<JsonNode> queryAwardList(Long strategyId);
 
-    public Response<JsonNode> getAward(Integer userId, Integer strategyId);
+    public Response<JsonNode> getAward(Long userId, Long strategyId);
 
 }
