+++ big-market-domain/src/main/java/app/xlog/ggbond/raffle/model/vo/DecrQueueVO.java
@@ -10,6 +10,6 @@ import lombok.NoArgsConstructor;
 @Getter
 @Builder
 public class DecrQueueVO {
-    private Integer strategyId;
-    private Integer awardId;
+    private Long strategyId;
+    private Long awardId;
 }
