+++ big-market-domain/src/main/java/app/xlog/ggbond/raffle/service/IRaffleService.java
@@ -5,7 +5,7 @@ import com.fasterxml.jackson.databind.node.ObjectNode;
 import java.util.List;
 
 public interface IRaffleService {
-    List<ObjectNode> queryAwardList(Integer strategyId);
+    List<ObjectNode> queryAwardList(Long strategyId);
 
-    Integer getAward(Integer userId, Integer strategyId);
+    Long getAward(Long userId, Long strategyId);
 }
