+++ big-market-domain/src/main/java/app/xlog/ggbond/raffle/service/armory/IRaffleArmory.java
@@ -5,17 +5,17 @@ package app.xlog.ggbond.raffle.service.armory;
  */
 public interface IRaffleArmory {
     // 传入一个策略ID，装配所有奖品，由于要还要拼接rule的名字，所以改成String格式
-    void assembleLotteryStrategyRuleCommon(Integer strategyId);
+    void assembleLotteryStrategyRuleCommon(Long strategyId);
 
     // 传入一个策略ID，装配除去锁出的四个奖品的其余奖品
-    void assembleLotteryStrategyRuleLock(Integer strategyId);
+    void assembleLotteryStrategyRuleLock(Long strategyId);
 
     // 传入一个策略ID，装配除去最后一个奖品的其余奖品
-    void assembleLotteryStrategyRuleLockLong(Integer strategyId);
+    void assembleLotteryStrategyRuleLockLong(Long strategyId);
 
     // 传入一个策略ID，装配大奖的抽奖池
-    void assembleLotteryStrategyRuleGrand(Integer strategyId);
+    void assembleLotteryStrategyRuleGrand(Long strategyId);
 
     // 传入一个策略ID，装配所有奖品的库存
-    void assembleLotteryStrategyAwardCount(Integer strategyId);
+    void assembleLotteryStrategyAwardCount(Long strategyId);
 }
