+++ big-market-domain/src/main/java/app/xlog/ggbond/user/service/IUserService.java
@@ -4,10 +4,10 @@ public interface IUserService {
     /**
      * 判断用户是否是黑名单用户
      **/
-    Boolean isBlacklistUser(Integer userId);
+    Boolean isBlacklistUser(Long userId);
 
     /**
      * 查询用户的抽奖次数
      **/
-    Integer queryRaffleTimesByUserId(Integer userId);
+    Long queryRaffleTimesByUserId(Long userId);
 }
