+++ big-market-infrastructure/src/main/java/app/xlog/ggbond/persistent/repository/StrategyRepository.java
@@ -8,5 +8,5 @@ import org.springframework.stereotype.Repository;
 @Repository
 public interface StrategyRepository extends JpaRepository<Strategy, Long> {
     @Query("select s from Strategy s where s.strategyId = ?1")
-    Strategy findByStrategyId(Integer strategyId);
+    Strategy findByStrategyId(Long strategyId);
 }
